# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API de Clientes y Direcciones
  description: Endpoints expuestos por `ClienteController`
  version: "1.0.0"
servers:
  - url: http://localhost:8080
tags:
  - name: Cliente
  - name: Direccion

paths:
  /cliente:
    get:
      tags: [Cliente]
      summary: Listar todos los clientes
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"

    post:
      tags: [Cliente]
      summary: Crear un nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cliente"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"

  /cliente/{idCliente}:
    parameters:
      - $ref: "#/components/parameters/IdCliente"
    get:
      tags: [Cliente]
      summary: Obtener un cliente por ID
      responses:
        "200":
          description: Datos del cliente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "400":
          description: Cliente inexistente
    put:
      tags: [Cliente]
      summary: Actualizar un cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cliente"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "400":
          description: ID del path y del body no coinciden
        "404":
          description: Cliente no encontrado
    delete:
      tags: [Cliente]
      summary: Eliminar un cliente
      responses:
        "200":
          description: Cliente eliminado
        "404":
          description: Cliente no encontrado

  /cliente/{idCliente}/direccion:
    parameters:
      - $ref: "#/components/parameters/IdCliente"
    get:
      tags: [Direccion]
      summary: Listar direcciones de un cliente
      responses:
        "200":
          description: Lista de direcciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Direccion"
        "400":
          description: Cliente inexistente
    post:
      tags: [Direccion]
      summary: Crear dirección para un cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Direccion"
      responses:
        "201":
          description: Dirección creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Direccion"
        "404":
          description: Cliente no encontrado

  /cliente/{idCliente}/direccionActual:
    parameters:
      - $ref: "#/components/parameters/IdCliente"
    get:
      tags: [Direccion]
      summary: Listar direcciones actuales de un cliente
      responses:
        "200":
          description: Lista de direcciones actuales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Direccion"
        "400":
          description: Cliente inexistente

  /cliente/{idCliente}/direccion/{idDireccion}:
    parameters:
      - $ref: "#/components/parameters/IdCliente"
      - $ref: "#/components/parameters/IdDireccion"
    put:
      tags: [Direccion]
      summary: Actualizar una dirección
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Direccion"
      responses:
        "200":
          description: Dirección actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Direccion"
        "404":
          description: Cliente o dirección no encontrados
    delete:
      tags: [Direccion]
      summary: Eliminar una dirección
      responses:
        "204":
          description: Dirección eliminada (sin contenido)
        "400":
          description: La dirección no pertenece al cliente
        "404":
          description: Cliente o dirección no encontrados

components:
  parameters:
    IdCliente:
      name: idCliente
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    IdDireccion:
      name: idDireccion
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

  schemas:
    Cliente:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          format: email
          example: juan@example.com
        telefono:
          type: string
          example: "+34 600 123 456"
        direcciones:
          type: array
          items:
            $ref: "#/components/schemas/Direccion"

    Direccion:
      type: object
      required: [id, direccion]
      properties:
        id:
          type: integer
          format: int64
          example: 10
        direccion:
          type: string
          example: C/ Mayor 12, 3º B
        municipio:
          type: string
          example: Madrid
        provincia:
          type: string
          example: Madrid
        cp:
          type: string
          example: "28013"
        esActual:
          type: boolean
          description: Dirección vigente
        esComprador:
          type: boolean
          description: Marca si la usa el comprador